requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
limits==4.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.71.0
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyperclip==1.9.0
pytest==8.3.5
pytest-flask==1.3.0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==78.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
Flask-Migrate==4.0.7